style: :art: Homepage design revision
Toolset section: Updated the list of the tools to add new ones and also to remove some. Contact us section: Changing the background and the style for the buttons(orange background and dark blue hover). Hero image section: Editing the software development card to use the gray background. Card reveal section: Changing the header color. Editing the illustrations to use the gray background. Footer: Reducing its height. Better alignment for the columns. Changing the style for the sections's orange divider. Contact us form: Fixing the margins for the button.  Making a standard for every button on the site according to width, height and icon. Also is necessary to use the orange secondary color background and primary color hover to see how it looks.

#25